sml: multiple changes
Updated hyperparameters to be more sane, training performance improves
greatly with these settings though more testing and tweaks are still
needed for optimal performance.

Randomized the amount of lives when starting an episode so models can
hopefully learn to operate differently when having 1 vs many lives left.

To emphasize the importance of the last life, a stronger punishment was
added for a game over. Movement rewards were also tweaked so moving
backwards is punished much less as it is sometimes necessary and useful
to have as an option. The punishment was fairly strict before so
previous models hardly ever considered doing it. The reward for moving
forward was also decreased but a new reward for level progress was
added. The goal with these changes is to incentivize staying on one life
for as long as possible, as dying and getting sent back will not reward
forward movement as much until the agent reaches where they died.